requirements.txt
numpy
opencv-python
matplotlib
scikit-image
imageio
torch
torchvision
torchsummary
tensorboardX
pyrealsense2; platform_system == "Windows" or platform_system == "Linux"
pyrealsense2-macosx; platform_system == "Darwin"
Pillow